fix(shutdown): retain child cancellation cleanup
