+++ app/main.py
@@ -522,7 +522,9 @@ async def lifespan(app: FastAPI):
         # still held; bridge teardown below may release those owner fences.
         if proxy_service is not None and hasattr(proxy_service, "drain_persistence_tasks"):
             try:
-                await proxy_service.drain_persistence_tasks(timeout_seconds=settings.shutdown_drain_timeout_seconds)
+                await proxy_service.drain_persistence_tasks(
+                    timeout_seconds=max(settings.shutdown_drain_timeout_seconds, 300.0)
+                )
             except Exception:
                 logger.warning("Failed to pre-drain proxy settlement tasks during shutdown", exc_info=True)
         if proxy_service is not None and hasattr(proxy_service, "mark_http_bridge_draining"):
